assignment_12_MCP/requirements.txt
# Canvas MCP Chatbot - Requirements
# Install with: pip install -r requirements.txt

# Core Chatbot
gradio>=4.0.0
openai>=1.0.0

# MCP Integration
mcp>=0.3.0
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0
httpx>=0.28.1
pydantic>=2.12.0

# Optional: For development
pytest>=7.0.0
black>=23.0.0
ruff>=0.1.0
